requirements.txt
fastapi
uvicorn
python-telegram-bot==20.5
gspread
oauth2client
python-dotenv
